+++ src/instrumentman/setup/app.py
@@ -1,22 +1,17 @@
 import os
 from logging import Logger, getLogger
 
-from click_extra import (
-    Choice,
-    prompt,
-    pause,
-    confirm,
-    echo
-)
+from rich.prompt import Prompt, Confirm, FloatPrompt
 from geocompy.communication import open_serial
 from geocompy.geo import GeoCom
 from geocompy.geo.gcdata import Prism
 from geocompy.geo.gctypes import GeoComCode
 
 from ..utils import (
-    echo_red,
-    echo_green,
-    echo_yellow
+    print_error,
+    print_success,
+    print_warning,
+    console
 )
 from ..targets import (
     TargetList,
@@ -32,17 +27,19 @@ def measure_targets(
     filepath: str
 ) -> TargetList | None:
     if os.path.exists(filepath):
-        action: str = prompt(
+        action: str = Prompt.ask(
             f"{filepath} already exists. Action",
+            console=console,
             default="replace",
-            type=Choice(["cancel", "replace", "append"])
+            case_sensitive=False,
+            choices=["cancel", "replace", "append"]
         )
         match action:
             case "cancel":
                 exit(0)
             case "append":
                 points = load_targets_from_json(filepath)
-                echo(f"Loaded targets: {points.get_target_names()}")
+                console.print(f"Loaded targets: {points.get_target_names()}")
             case _:
                 points = TargetList()
     else:
@@ -50,10 +47,10 @@ def measure_targets(
 
     logger.info("Starting target measurements")
     ptid: str
-    while ptid := prompt("Point ID (or nothing to finish)", type=str):
+    while ptid := Prompt.ask("Point ID (or nothing to finish)"):
         if ptid in points:
-            remove = confirm(
-                f"{ptid} already exists. Overwrite?"
+            remove = Confirm.ask(
+                f"Overwrite already existing {ptid}"
             )
             if remove:
                 points.pop_target(ptid)
@@ -62,7 +59,7 @@ def measure_targets(
         logger.info(f"Recording {ptid}")
         resp_target = tps.bap.get_prism_type()
         if resp_target.params is None:
-            echo_yellow("Could not retrieve target type.")
+            print_warning("Could not retrieve target type.")
             logger.error(
                 f"Could not retrieve target type, skipping ({resp_target})"
             )
@@ -70,28 +67,30 @@ def measure_targets(
 
         target = resp_target.params
 
-        user_target: str = prompt(
+        user_target: str = Prompt.ask(
             "Prism type",
+            console=console,
             default=target.name,
-            type=Choice([e.name for e in Prism if e.name != 'USER'])
+            case_sensitive=False,
+            choices=[e.name for e in Prism if e.name != 'USER']
         )
         if target != Prism[user_target]:
             resp_settarget = tps.bap.set_prism_type(user_target)
             if resp_settarget.error != GeoComCode.OK:
-                echo_yellow("Could not update prism type")
+                print_warning("Could not update prism type")
                 logger.error(
                     f"Could not update prism type, skipping ({resp_settarget})"
                 )
                 continue
             else:
-                echo(f"Updated prism type to {user_target}")
+                console.print(f"Updated prism type to {user_target}")
                 logger.info(
                     f"Updated prism type to {user_target} from user input"
                 )
         target = Prism[user_target]
 
         if target == Prism.USER:
-            echo_yellow(
+            print_warning(
                 "User defined prisms are currently not supported."
             )
             logger.error(
@@ -102,37 +101,37 @@ def measure_targets(
         resp_height = tps.tmc.get_target_height()
         height = resp_height.params
         if height is None:
-            echo_yellow("Could not retrieve target height.")
+            print_warning("Could not retrieve target height.")
             logger.error(
                 f"Could not retrieve target height, skipping ({resp_height})"
             )
             continue
 
-        user_height: float = prompt(
+        user_height: float = FloatPrompt.ask(
             "Target height",
-            default=f"{resp_height.params:.4f}",
-            type=float
+            console=console,
+            default=height
         )
         resp_setheight = tps.tmc.set_target_height(user_height)
         if resp_setheight.error != GeoComCode.OK:
-            echo_yellow("Could not update target height")
+            print_warning("Could not update target height")
             logger.error(
                 f"Could not update target height, skipping ({resp_setheight})"
             )
             continue
         else:
-            echo(f"Updated target height to {user_height:.4f}")
+            console.print(f"Updated target height to {user_height:.4f}")
             logger.info(
                 f"Updated target height to {user_height:.4f} from user input"
             )
 
-        pause("Aim at target, then press any key...")
+        console.input("Aim at target, then press ENTER...")
 
         tps.aut.fine_adjust(0.5, 0.5)
         tps.tmc.do_measurement()
         resp = tps.tmc.get_simple_coordinate(10)
         if resp.params is None:
-            echo_yellow("Could not measure target")
+            print_warning("Could not measure target")
             logger.error("Could not measure target, skipping")
             continue
 
@@ -145,12 +144,16 @@ def measure_targets(
             )
         )
 
-        echo_green(f"{ptid} recorded")
+        print_success(f"{ptid} recorded")
         logger.info(f"{ptid} recorded")
-        if not confirm("Record more targets?", default=True):
+        if not Confirm.ask(
+            "Record more targets",
+            console=console,
+            default=True
+        ):
             break
 
-    echo_green("Target measurement finished")
+    print_success("Target measurement finished")
     logger.info("Target measurement finished")
 
     return points
@@ -176,11 +179,11 @@ def main_measure(
         tps = GeoCom(com, logger.getChild("instrument"))
         targets = measure_targets(tps, logger, output)
         if targets is None:
-            echo_red("Program was cancelled or no targets were recorded")
+            print_error("Program was cancelled or no targets were recorded")
             logger.info("Program was cancelled or no targets were recorded")
             exit(0)
 
     if targets is not None:
         export_targets_to_json(output, targets)
-        echo_green(f"Saved target results to '{output}'")
+        print_success(f"Saved target results to '{output}'")
         logger.info(f"Saved target results to '{output}'")